requirements.txt
streamlit
google.generativeai
os
sys
random